refactor(app): split status summary sections, restore artifact open capability
Decompose session-status-summary into shell + git + artifact files. Per-row
open / reveal capability now follows the old Files-tab pattern: panel stats
artifact paths via platform.statPaths, disables actions when the file no
longer exists or the host can't open paths, and surfaces openPath /
showItemInFolder / openLink failures through a toast instead of swallowing
them. Worktree row applies the same canOpen / disabled treatment.

snap target now asserts exact seeded counts (4 todos, 2 artifacts, 2
sources) and waits for per-path diff stats to render before screenshotting,
so partial-seed regressions and silent normalizeArtifactPathKey breakage
cannot pass.
