+++ packages/app/src/components/session/session-status-panel.tsx
@@ -1,7 +1,11 @@
-import { createMemo, type Accessor } from "solid-js"
+import { createMemo, createResource, type Accessor } from "solid-js"
 import { useParams } from "@solidjs/router"
+import { showToast } from "@opencode-ai/ui/toast"
 import type { Part } from "@opencode-ai/sdk/v2"
 import { useGlobalSync } from "@/context/global-sync"
+import { useLanguage } from "@/context/language"
+import { canOpenLocalPath, usePlatform } from "@/context/platform"
+import { useServer } from "@/context/server"
 import { useSync } from "@/context/sync"
 import { normalizeArtifactPathKey, type FilesTabEntry } from "@/pages/session/files-tab-state"
 import { aggregateFiles } from "@/pages/session/session-aggregate-files"
@@ -15,6 +19,9 @@ export function SessionStatusPanel(props: {
   const params = useParams()
   const globalSync = useGlobalSync()
   const sync = useSync()
+  const language = useLanguage()
+  const platform = usePlatform()
+  const server = useServer()
 
   const parts = createMemo<Part[]>(() => {
     if (!params.id) return []
@@ -61,6 +68,65 @@ export function SessionStatusPanel(props: {
     return map
   })
 
+  // Stat artifact files so per-row open/reveal buttons can disable themselves
+  // when the file is gone (deleted between the agent writing it and the user
+  // clicking). Mirrors the old Files-tab stat resource. On web (no statPaths
+  // capability) every file is treated as existing — clicking still no-ops
+  // because canOpenLocalPath returns false there.
+  const artifactPaths = createMemo(() => (props.artifactFiles?.() ?? []).map((file) => file.path))
+  const [artifactStats] = createResource(artifactPaths, async (paths) => {
+    if (paths.length === 0) return {} as Record<string, { size: number; exists: boolean }>
+    if (!platform.statPaths) {
+      return Object.fromEntries(paths.map((path) => [path, { size: 0, exists: true }]))
+    }
+    return platform.statPaths(paths)
+  })
+  // While the resource is pending the row should still feel actionable, so
+  // assume the file exists until we know otherwise. This matches how the row
+  // first renders right after a turn — the file did exist when the agent wrote
+  // it, and the click is no-op-safe even if stat later reports missing.
+  const artifactExists = (path: string) => artifactStats()?.[path]?.exists ?? true
+
+  const reportFailure = (error: unknown) => {
+    showToast({
+      variant: "error",
+      title: language.t("common.requestFailed"),
+      description: error instanceof Error ? error.message : String(error),
+    })
+  }
+
+  const canOpenWorktreeDirectory = (directory: string): boolean =>
+    !!(canOpenLocalPath(platform) && server.isLocal() && platform.openPath && directory)
+  const openWorktreeDirectory = (directory: string) => {
+    if (!canOpenWorktreeDirectory(directory) || !platform.openPath) return
+    void platform.openPath(directory).catch(reportFailure)
+  }
+
+  const canOpenArtifactFile = (path: string): boolean =>
+    !!(canOpenLocalPath(platform) && server.isLocal() && platform.openPath && artifactExists(path))
+  const openArtifactFile = (path: string) => {
+    if (!canOpenArtifactFile(path) || !platform.openPath) return
+    void platform.openPath(path).catch(reportFailure)
+  }
+
+  const canRevealArtifactFile = (path: string): boolean =>
+    !!(canOpenLocalPath(platform) && server.isLocal() && platform.showItemInFolder && artifactExists(path))
+  const revealArtifactFile = (path: string) => {
+    if (!canRevealArtifactFile(path) || !platform.showItemInFolder) return
+    void platform.showItemInFolder(path).catch(reportFailure)
+  }
+
+  // openLink is the only mandatory Platform method, so there is no capability
+  // gate. It may still throw synchronously (e.g. an invalid scheme) — catch
+  // and surface the same failure toast as the file/directory openers.
+  const openSourceLink = (url: string) => {
+    try {
+      platform.openLink(url)
+    } catch (error) {
+      reportFailure(error)
+    }
+  }
+
   return (
     <div class="h-full min-h-0 overflow-y-auto">
       <SessionStatusSummary
@@ -73,7 +139,14 @@ export function SessionStatusPanel(props: {
         diffStats={diffStats}
         artifactFiles={props.artifactFiles}
         diffsByPath={diffsByPath}
+        canOpenWorktreeDirectory={canOpenWorktreeDirectory}
+        canOpenArtifactFile={canOpenArtifactFile}
+        canRevealArtifactFile={canRevealArtifactFile}
         onNavigateReview={props.onNavigateReview}
+        onOpenWorktreeDirectory={openWorktreeDirectory}
+        onOpenArtifactFile={openArtifactFile}
+        onRevealArtifactFile={revealArtifactFile}
+        onOpenSourceLink={openSourceLink}
       />
     </div>
   )
